+++ planning/behavior_path_planner/include/behavior_path_planner/scene_module/pull_out/pull_out_module.hpp
@@ -15,13 +15,13 @@
 #ifndef BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OUT__PULL_OUT_MODULE_HPP_
 #define BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OUT__PULL_OUT_MODULE_HPP_
 
-#include "behavior_path_planner/scene_module/pull_out/geometric_pull_out.hpp"
-#include "behavior_path_planner/scene_module/pull_out/pull_out_parameters.hpp"
-#include "behavior_path_planner/scene_module/pull_out/pull_out_path.hpp"
-#include "behavior_path_planner/scene_module/pull_out/shift_pull_out.hpp"
 #include "behavior_path_planner/scene_module/scene_module_interface.hpp"
 #include "behavior_path_planner/util/geometric_parallel_parking/geometric_parallel_parking.hpp"
 #include "behavior_path_planner/util/path_shifter/path_shifter.hpp"
+#include "behavior_path_planner/util/pull_out/geometric_pull_out.hpp"
+#include "behavior_path_planner/util/pull_out/pull_out_parameters.hpp"
+#include "behavior_path_planner/util/pull_out/pull_out_path.hpp"
+#include "behavior_path_planner/util/pull_out/shift_pull_out.hpp"
 
 #include <lane_departure_checker/lane_departure_checker.hpp>
 #include <lanelet2_extension/utility/message_conversion.hpp>
