+++ planning/behavior_path_planner/include/behavior_path_planner/util/pull_out/geometric_pull_out.hpp
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OUT__GEOMETRIC_PULL_OUT_HPP_
-#define BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OUT__GEOMETRIC_PULL_OUT_HPP_
+#ifndef BEHAVIOR_PATH_PLANNER__UTIL__PULL_OUT__GEOMETRIC_PULL_OUT_HPP_
+#define BEHAVIOR_PATH_PLANNER__UTIL__PULL_OUT__GEOMETRIC_PULL_OUT_HPP_
 
-#include "behavior_path_planner/scene_module/pull_out/pull_out_path.hpp"
-#include "behavior_path_planner/scene_module/pull_out/pull_out_planner_base.hpp"
 #include "behavior_path_planner/util/geometric_parallel_parking/geometric_parallel_parking.hpp"
+#include "behavior_path_planner/util/pull_out/pull_out_path.hpp"
+#include "behavior_path_planner/util/pull_out/pull_out_planner_base.hpp"
 
 #include <autoware_auto_planning_msgs/msg/path_with_lane_id.hpp>
 
@@ -38,4 +38,4 @@ public:
 };
 }  // namespace behavior_path_planner
 
-#endif  // BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OUT__GEOMETRIC_PULL_OUT_HPP_
+#endif  // BEHAVIOR_PATH_PLANNER__UTIL__PULL_OUT__GEOMETRIC_PULL_OUT_HPP_
