+++ planning/behavior_path_planner/include/behavior_path_planner/util/pull_over/geometric_pull_over.hpp
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OVER__GEOMETRIC_PULL_OVER_HPP_
-#define BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OVER__GEOMETRIC_PULL_OVER_HPP_
+#ifndef BEHAVIOR_PATH_PLANNER__UTIL__PULL_OVER__GEOMETRIC_PULL_OVER_HPP_
+#define BEHAVIOR_PATH_PLANNER__UTIL__PULL_OVER__GEOMETRIC_PULL_OVER_HPP_
 
-#include "behavior_path_planner/scene_module/pull_over/pull_over_planner_base.hpp"
 #include "behavior_path_planner/util/geometric_parallel_parking/geometric_parallel_parking.hpp"
 #include "behavior_path_planner/util/occupancy_grid_based_collision_detector/occupancy_grid_based_collision_detector.hpp"
+#include "behavior_path_planner/util/pull_over/pull_over_planner_base.hpp"
 
 #include <lane_departure_checker/lane_departure_checker.hpp>
 
@@ -69,4 +69,4 @@ protected:
 };
 }  // namespace behavior_path_planner
 
-#endif  // BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OVER__GEOMETRIC_PULL_OVER_HPP_
+#endif  // BEHAVIOR_PATH_PLANNER__UTIL__PULL_OVER__GEOMETRIC_PULL_OVER_HPP_
