+++ planning/behavior_path_planner/include/behavior_path_planner/util/pull_over/goal_searcher_base.hpp
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OVER__GOAL_SEARCHER_BASE_HPP_
-#define BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OVER__GOAL_SEARCHER_BASE_HPP_
+#ifndef BEHAVIOR_PATH_PLANNER__UTIL__PULL_OVER__GOAL_SEARCHER_BASE_HPP_
+#define BEHAVIOR_PATH_PLANNER__UTIL__PULL_OVER__GOAL_SEARCHER_BASE_HPP_
 
 #include "behavior_path_planner/data_manager.hpp"
 #include "behavior_path_planner/parameters.hpp"
-#include "behavior_path_planner/scene_module/pull_over/pull_over_parameters.hpp"
+#include "behavior_path_planner/util/pull_over/pull_over_parameters.hpp"
 
 #include <geometry_msgs/msg/pose_stamped.hpp>
 
@@ -73,4 +73,4 @@ protected:
 };
 }  // namespace behavior_path_planner
 
-#endif  // BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OVER__GOAL_SEARCHER_BASE_HPP_
+#endif  // BEHAVIOR_PATH_PLANNER__UTIL__PULL_OVER__GOAL_SEARCHER_BASE_HPP_
