+++ planning/behavior_path_planner/src/util/pull_out/geometric_pull_out.cpp
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "behavior_path_planner/scene_module/pull_out/geometric_pull_out.hpp"
+#include "behavior_path_planner/util/pull_out/geometric_pull_out.hpp"
 
-#include "behavior_path_planner/scene_module/pull_out/util.hpp"
+#include "behavior_path_planner/util/pull_out/util.hpp"
+#include "behavior_path_planner/utilities.hpp"
+
+#include <lanelet2_extension/utility/utilities.hpp>
 
 using lanelet::utils::getArcCoordinates;
 using motion_utils::findNearestIndex;
